refactor(ai): split cross-specialty rules, extract NSAID pattern, fix Kawasaki comment, dedup thiazide-hypoK
- Split the four CROSS-* contraindication rules (NSAID+CHF, STATIN+HEPATIC,
  ACE/ARB+PREG, METFORMIN+GFR) out of cross-specialty.ts into a dedicated
  contraindications.ts module. cross-specialty.ts now focuses on multi-drug
  risk patterns (triple-whammy AKI, QT+hypoK, thiazide+hypoK, CHEMO-*,
  ONCO-VTE-NEURO, obstetric teratogens, hepatic/renal hepatotoxin combos).
  Wired checkContraindications into the review-service pipeline alongside
  checkCrossSpecialtyPatterns. Tests moved to contraindications-rules.test.ts.
- Extracted the duplicated NSAID_PATTERN from cross-specialty.ts and
  age-stratified.ts into shared-drug-patterns.ts (joins METFORMIN_PATTERN
  there). Union pattern — broader of the two prior definitions — is used
  everywhere, so no rule loses matches.
- Fixed the stale PEDI-ASPIRIN-VIRAL-001 comment in age-stratified.ts that
  claimed the rule fires on Kawasaki-alone to surface documentation; the
  rule requires concurrent viral illness evidence and intentionally does
  NOT fire on Kawasaki-only (confirmed by the existing test).
- Extended consolidateRuleFlags to suppress CRITICAL-LAB-POTASSIUM when
  CROSS-THIAZIDE-HYPOK-001 fires (parallels the existing CROSS-QT-HYPOK-001
  dedup from #854). When both cross-specialty hypoK rules fire together the
  thiazide and QT flags survive — only the generic critical-value flag is
  dropped. Updated the stale "overlap is deliberate and left un-deduplicated"
  comment on CROSS-THIAZIDE-HYPOK-001 to point at the new dedup path.

Closes #902
Closes #903
Closes #904
Closes #878
